Get details of a spot by id fix

+++ backend/db/migrations/20220816221226-create-spot.js
@@ -24,10 +24,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       name: {
         type: Sequelize.STRING
